# docs/uapoly-openapi.yaml
openapi: 3.1.0
...
info:
  title: UApoly
  version: '1.0'
  description: The API specification for interacting with the UApoly backend.
  x-logo:
    url: /docs/Logo.webp
servers:
  - url: 'http://localhost:{port}'
    description: Local instance
    variables:
      port:
        default: '3000'
        description: The port where the local instance of the UApoly backend is running.
    x-last-modified: 1680690602604
paths:
...
  /friend:
    summary: Get the user's friend list
    description: Get the friend list associated with the currently logged-in user.
    get:
      tags: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendList'
          description: A JSON array containing the friend list.
          x-last-modified: 1680688989769
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                message: Account not found
          description: Account not found
          x-last-modified: 1680691111163
      operationId: /friend
      summary: Get the user's friend list
      description: Get the friend list associated with the currently logged-in user.
    x-last-modified: 1680688852703
...
  /game/list:
    post:
      summary: Get the current games
      operationId: /game/list
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameListPage'
              examples:
                Example:
                  value:
                    pageCount: 1
                    games:
                      - id: 2
                        players: 1
                        maxPlayers: 2147483647
                        salary: 200
                        initialMoney: 2500
                        private: true
                        started: false
                        friendsOnly: false
                        name: My Game
                      - id: 3
                        players: 1
                        maxPlayers: 2147483647
                        salary: 200
                        initialMoney: 2500
                        private: false
                        started: false
                        friendsOnly: false
                        name: Our Game
        '400':
          description: Missing argument or invalid page number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example:
                  value:
                    message: Invalid page number
      description: Get the current games
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                page:
                  type: integer
                  minimum: 1
                  description: The requested page
              required:
                - page
        description: 'A JSON object with :'
  /game/create:
    post:
      summary: Create a new game
      operationId: /game/create
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameMetadata'
              examples:
                Example:
                  value:
                    id: 1
                    players: 1
                    maxPlayers: 2
                    salary: 0
                    initialMoney: 1
                    private: true
                    started: true
                    friendsOnly: true
                    name: Their Game
        '400':
          description: Missing or invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example 1:
                  value:
                    message: Invalid initial money
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example:
                  value:
                    message: User not found
      description: Create a new game
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameCreationMetadata'
            examples:
              Example:
                value:
                  maxPlayers: 2
                  salary: 0
                  initialMoney: 1
                  friendsOnly: true
                  password: pa$$word
                  locale: en-US
                  name: A Game
        description: 'A JSON object with :'
  /game/join:
    post:
      summary: Join a game
      operationId: /game/join
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example:
                  value:
                    message: Joined game
        '400':
          description: Missing or invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example:
                  value:
                    message: This game has already started
        '401':
          description: Missing password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example:
                  value:
                    message: Missing password
        '403':
          description: 'Invalid password, or not friends with the game master'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example:
                  value:
                    message: You are not friends with the game master
        '404':
          description: Game or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example:
                  value:
                    message: Game not found
      description: Join a game
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                gameId:
                  type: integer
                  description: The ID of the game to join
                  minimum: 1
                password:
                  type: string
                  format: password
                  description: 'The password of the game, if the game is private'
              required:
                - gameId
        description: 'A JSON object with :'
...
  /game/search:
    post:
      summary: Search for a game
      operationId: /game/search
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameMetadata'
              examples:
                Example:
                  value:
                    - id: 1
                      name: My UApoly Game
                      players: 1
                      maxPlayers: 2147483647
                      salary: 200
                      initialMoney: 2500
                      private: true
                      started: false
                      friendsOnly: false
                    - id: 2
                      name: Our UApoly Game
                      players: 1
                      maxPlayers: 2147483647
                      salary: 200
                      initialMoney: 2500
                      private: true
                      started: false
                      friendsOnly: false
                    - id: 3
                      name: Their UApoly Game
                      players: 1
                      maxPlayers: 2147483647
                      salary: 200
                      initialMoney: 2500
                      private: true
                      started: false
                      friendsOnly: false
                    - id: 4
                      name: This Is Not A UApoly Game
                      players: 1
                      maxPlayers: 2147483647
                      salary: 200
                      initialMoney: 2500
                      private: true
                      started: false
                      friendsOnly: false
                    - id: 5
                      name: Winners only
                      players: 1
                      maxPlayers: 2147483647
                      salary: 200
                      initialMoney: 2500
                      private: true
                      started: false
                      friendsOnly: false
                    - id: 6
                      name: Pro Players
                      players: 1
                      maxPlayers: 2147483647
                      salary: 200
                      initialMoney: 2500
                      private: true
                      started: false
                      friendsOnly: false
                    - id: 7
                      name: UApoly Game
                      players: 1
                      maxPlayers: 2147483647
                      salary: 200
                      initialMoney: 2500
                      private: true
                      started: false
                      friendsOnly: false
        '400':
          description: Invalid arguments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                Example:
                  value:
                    message: Invalid page
      description: Search for created lobbies
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the lobby to search for
                  minLength: 1
                page:
                  type: integer
                  minimum: 1
                  description: The requested page of results
              required:
                - name
            examples:
              Example:
                value:
                  name: UApoly Game
                  page: 1
        description: 'A JSON object with :'
components:
  schemas:
    Friend:
      description: Represents a friend request.
      required:
        - firstAccountLogin
        - secondAccountLogin
        - accepted
      type: object
      properties:
        firstAccountLogin:
          description: The name of the sender account
          type: string
        secondAccountLogin:
          description: The name of the receiver account
          type: string
        accepted:
          description: 'True if the request is accepted, false if it''s pending.'
          type: boolean
      x-last-modified: 1680688756160
    FriendList:
      description: A JSON array containing a friend list.
      type: array
      items:
        $ref: '#/components/schemas/Friend'
        description: A friend
      x-last-modified: 1680689436014
...
    Message:
      description: 'Represents an message, either a success or an error'
      required:
        - message
      type: object
      properties:
        message:
          description: The message
          type: string
      x-last-modified: 1680692391186
...
    GameMetadata:
      type: object
      x-examples:
        Example 1:
          id: 2
          players: 1
          maxPlayers: 2147483647
          salary: 200
          initialMoney: 2500
          private: true
          started: false
          friendsOnly: false
      description: Describe a game's metadata
      properties:
        id:
          type: integer
          minimum: 1
          description: The game ID
        players:
          type: integer
          minimum: 1
          description: The amount of players currently in the game
        maxPlayers:
          type: integer
          minimum: 2
          description: The maximum amount of players
          maximum: 2147483647
        salary:
          type: integer
          description: The amout of money a player will get by passing the start slot
        initialMoney:
          type: integer
          minimum: 1
          description: The amount of money the players will start with
        private:
          type: boolean
          description: 'True if the game is password-protected, false otherwise'
        started:
          type: boolean
          description: 'True if the game has already started, false otherwise'
        friendsOnly:
          type: boolean
          description: 'True if you need to be friends with the game master to join, false otherwise'
        name:
          type: string
          description: The name of the lobby
      required:
        - id
        - players
        - maxPlayers
        - salary
        - initialMoney
        - private
        - started
        - friendsOnly
        - name
    GameCreationMetadata:
      type: object
      x-examples:
        Example 1:
          maxPlayers: 4
          salary: 200
          initialMoney: 2500
          password: password
          friendsOnly: false
      description: Describe a game's metadata
      x-stoplight:
        id: 049f57577e1e5
      examples:
        - maxPlayers: 8
          salary: 200
          initialMoney: 2500
          friendsOnly: true
          password: pa$$word
          locale: en-US
      properties:
        maxPlayers:
          type: integer
          minimum: 2
          description: The maximum amount of players
          maximum: 2147483647
        salary:
          type: integer
          description: The amout of money a player will get by passing the start slot
        initialMoney:
          type: integer
          minimum: 1
          description: The amount of money the players will start with
        friendsOnly:
          type: boolean
          description: 'True if you need to be friends with the game master to join, false otherwise'
        password:
          type: string
          format: password
          description: The password of the game. If not set the game will be public.
        locale:
          type: string
          pattern: '[a-z]+-[A-Z]+'
          description: The locale of the board
        name:
          type: string
          description: The name of the lobby
      required:
        - salary
        - initialMoney
        - friendsOnly
        - locale
        - name
    GameListPage:
      type: object
      x-examples:
        Example 1:
          pageCount: 1
          games:
            - id: 2
              players: 1
              maxPlayers: 2147483647
              salary: 200
              initialMoney: 2500
              private: true
              started: false
              friendsOnly: false
            - id: 3
              players: 1
              maxPlayers: 2147483647
              salary: 200
              initialMoney: 2500
              private: false
              started: false
              friendsOnly: false
      description: A page of games
      properties:
        pageCount:
          type: integer
          minimum: 1
          description: The total page count
        games:
          type: array
          description: The games currently on the server
          items:
            $ref: '#/components/schemas/GameMetadata'
      required:
        - pageCount
        - games
...
  securitySchemes:
    JWTAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: |-
        Example: 
        > Authorization: Bearer <token>
      x-last-modified: 1680684271664
...
security:
  - JWTAuth: []
...